int -> NSInteger transition.
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/apps/gorm/trunk@37359 72102866-910b-0410-8b05-ffd578937521

+++ GormCore/GormBoxEditor.m
@@ -99,8 +99,8 @@
 
 - (void) deleteSelection
 {
-  int i = 0;
-  int count = [selection count];
+  NSInteger i = 0;
+  NSInteger count = [selection count];
   id temp = nil;
   
   for (i = count - 1; i >= 0; i--)
